Applied Andy Herkey's ipmodem password patch [round 1]
Comments from his email:

Attached are the rough patches to get the IPmodem to work with a
password. These patches work on my 8703e, but I can't test it without a
password to see if I broke anything.

I added a couple more signals to pppob.cc and a Close method to
m_ipmodem.cc, so the modem connection is gracefully terminated when pppd
is shutdown. I don't know if the close sequence will work on any other
devices though.