Applied Andy Herkey's ipmodem password patch [round 2]
Comments from his email:

Here are the latest patches for the ipmodem. These files have been
merged with the latest cvs.
I added a check in the open function to see if the modem will respond to
commands.
I removed the m_filter from the Write function as Rick suggested and it
works great for me.