Fixed RFC 2426 conformance bug
RFC 2426, 3.1.1 states that FN MUST be present in the vcard object.
If there is no name available, then we use the Company name if available.
It should be, since the Blackberry requires either name or Company in
order to create a new address.  If Company is blank too, then we insert
"(Blank Name)".

On the flip side of this same bug, it is possible to receive
non-conformant vcard data from evolution / opensync that does not include
FN.  Since the Blackberry doesn't require this information, remove the
exception, and only throw an exception if FN and Company are both blank.