Birthday field and sync support
From Lee Dixon's email:

Attached is a patch file, and the associated files that were changed, to
support birthday in the Barry::Contact and for syncing in the opensync
plugin.

Regarding OpenSync... it seems that adding support for a new field does
not automatically cause that new data to be sync'd... makes sense
because it doesn't appear as though the record changed.

To see the birthdays I had in evo-2 show up, I had to clear my BB
contacts and the opensync group and start over to get the birthdays
transferred.