Added support for multiple work/home phone numbers to the opensync plugin
Thanks to Richard Esplin for his test data.

This fix has a conflict when a cell phone number is marked as the
preferred contact as well.
See the following URL for the full details (also pasted below)
http://sourceforge.net/mailarchive/message.php?msg_name=20081030231045.GA28989%40foursquare.net

This fix also changes the VCARD output... pager phone numbers are
given the TYPE code of "pager" instead of "msg".

Added OtherPhone field as well.

Description of field type conflict:

Thanks very much, Richard, for sending your test data!  Sorry for the
delay in getting back to you.  I've been busy with the Ontario Linux Fest.

I've taken a close look at your data, and at the plugin code, and have
fixed things up.  The handling of multiple phone numbers is much more
robust now than before, as well as handling of VCARD parameters.  Some
parameter values were missed before.

I've run into a conflict between data sets, and would like your feedback
as well as the feedback of the list.

The Blackberry has the following phone number fields that Barry can parse:

	Phone
	Fax
	WorkPhone
	HomePhone
	MobilePhone
	Pager
	WorkPhone2
	HomePhone2
	OtherPhone

Of this list, I assume that "Phone" is the preferred Blackberry phone number,
and I tag this accordingly in the VCARD data.  For example, if Phone held
the number 519-555-1212, then during the sync, Barry would output:

TEL;TYPE=pref:519-555-1212

The problem comes in when the number is a cell phone number and also marked as
the preferred number.  For example, if you set the cell phone number in your
KAddresBook settings as the preferred contact number, this is what I get
as input to the plugin:

TEL;TYPE=cell;TYPE=pref:519-555-1212

In the most recent code which I'm about to commit to CVS, it will take
that number and insert it into the MobilePhone field and the Phone field.
On the next sync, Barry will send this to KAddressBook:

TEL;TYPE=pref:519-555-1212
TEL;TYPE=cell:519-555-1212

I'm not sure yet how KAddressBook will handle this.  I'm testing with
Evolution.

There is no way that I can mark the MobilePhone field as "preferred" in
the Blackberry, so if I don't duplicate the data, the preferred status
may be lost.

Comments are welcome.

- Chris