Large parser/builder API change, adding IConverter support
This change starts at the top (Mode::Desktop) which takes an optional
IConverter object for parsing and building conversions.  The Desktop
class passes this IConverter object to Packet, Parser, and Builder
(and thereby, the corresponding Record classes) as needed.  This adds
an extra pointer to a fair number of calls in the stack, but maintains
flexibility without putting too much of a burden on Parser writers,
which could be application writers.

The primary use of this architecture is to pass in an IConverter
configured to convert to UTF-8... this is basically the only useful
conversion method, but others are possible if you want to play.