ServiceBook: added IConverter support, and general cleanup
Refactored ServiceBook class and added some experimental template
functions to generalize the FieldLink table processing.  If possible,
this needs to be expanded to the general record classes, to reduce
the amount of code duplication during processing of FieldLinks and
CommonField structures.

Also, ServiceBook deals with packed fields, which should also be
extrapolated for general use, and reused if possible for the new
"Calendar - All" Blackberry databases that are coming out on the new
devices and firmare.