Removed cache code in favour of using opensync's hashtable
API changes in opensync 0.4x have removed access to the .opensync
config directories, so make use of the hashtable in a bit of a
hackish way, similar to how the cache was used.

The cache simply kept track of what record ID's existed in the
device on the latest sync.  Since the Blackberry keeps track
of dirty flags for us, we don't need to actually hash all the
device data and load it over USB in order to find changes,
but this doesn't directly jive with the hashtable API.

We use a counter, and then during getchanges, calculate the
record's hash as:

	(last_hash + (dirty ? 1 : 0))