Fixed include errors from Nicolas Vivien's tree
Some headers are platform specific, and must never be installed,
since applications that make use of Barry should not need to care
what platform they are running on.

If your app needs to know endian settings, make it determine it
on its own.  And if the endian setting matters to anything related
to interfacing with the Blackberry, move it into the library, and
abstract it away.