desktop: add one-way locking mechanism so OpenSync40 always loads first
Opensync 0.22 has bugs in the way it loads its modules,
such as plugins, etc.  It doesn't load the plugins using
RTLD_LOCAL, and so libopensync.so.0 is loaded indirectly since
it is a dependency of the plugin itself, and pollutes the
symbol table, causing symbol clashes when 0.40 is loaded.

We need to make sure that OpenSync40 is loaded first if
desired.  This patch checks whether OpenSync22 was
loaded first, and if so, causes a std::logic_error exception
in the OpenSync40 constructor.