gui: added actual == total check to backup code
Added code to check that the actual number of backed up records matches
the total number available, as reported by the device, and display a
warning to the user if they don't match.

It is suspected that some international characters in the record data
cause the BlackBerry to use a different low-level protocol for those
records.  This protocol is not yet supported by Barry.