desktop: removed msynctool and osynctool source code from API classes
Copyright note: the API classes are closely based on the msynctool
and osynctool examples, but mostly have to, due to the nature of the API.

English status and error messages were sometimes copied verbatim,
as well as a simple message translation function.

It is my view that, due to the work added to make this a flexible
dlopen() API, and the porting to C++, and due to the fact that the
opensync API is fixed so that there is usually only one way to do
the basic operations, that this work is likely unique enough to
be considered copyrighted by Net Direct.