desktop: fixed strange segfault bug when client dies
When the TCP client is killed, the server segfaults, due to some
odd behaviour.  It is unknown exactly why this segfault occurs,
but it is likely due to the connections being deleted (via 'delete this')
before the final event is processed.

I *think* it is a bug in wxWidgets, but I was unable to track it
down.

Instead, I fixed it by overriding the 'delete this' behaviour,
removing it, and letting the container in SyncStatusDlg deleting
all the connections on close instead.