desktop: fixed auto_ptr<> forward declaration warnings
When a class uses an auto_ptr<> with a forward declaration, and then
does not have a destructor defined, then all classes that use
that class need to have the full definition of the forward
declaration, not just the .cc module of the class.

To avoid that, defined some destructors, to confine the destructor
code to one place, and therefore the header dependencies as well.