desktop: major rework of GroupCfgDlg and supporting code
Reworked GroupCfgDlg to keep existing plugin data if at all possible.
It now stores a nested map of engine, application, and plugin data.

	map<string, plugin_ptr> appmap
	map<engine*, appmap> engmap

With this nested map, plugins are stored as the user fiddles with the
controls, never deleting his settings until the dialog finally closes.
So, the original plugin_ptr config from the opensync config, that
has the member_id, will be kept, and re-used, if the user makes a
simple config change.

On the calling side in Sync_Mode.cc, the code has been changed to
clean up after itself if the engine was changed.  Also, the state
of the existing and new opensync configs are compared, and the user
is warned if a 1-Way reset is required due to his changes, and given
the chance to abort.  Finally the exception handling has been cleaned
up, since all the major decisions are now made before doing the save,
and the save should rarely fail now.