lib: added LDIF mapping support for Contact::Image fields
Note that jpegPhoto looks like the only LDIF field for photo images... it
is unknown which format will come from the BlackBerry in the Image field,
so we can't guarantee that "Image" will be in JPG format.

Therefore the LDIF mapping of "jpegPhoto" to "Image" is not done
automatically.

The LDIF mapping can be done manually from the btool command line
with "-m jpegPhoto,Image,Image" and hope that your LDAP software
and your BlackBerry uses JPG.

Reading photos from LDIF should be fine, since the BlackBerry
seems to handle most formats, when going in the LDIF -> BlackBerry
direction.