desktop: Mode_Sync: disconnect members when saving a new group
Not all code paths disconnected group members before saving a new group.
This bug showed up most prominently when changing engines for a given
config.

Also in this patch, the cleanup of the old engine's group is done after
the saving of the new one, so as not to delete something we may need
later.  This might cause problems though, if the save gets half-done,
and the next detection cannot complete.