libsync: moved vformat-related code from os22 to library
Giant patch, which moves the vbase and vformat related code out of
the opensync 0.22 plugin and into the library, creating a new
optional libbarrysync for this code, which can be hopefully shared by both
plugins.

This is step 1 of the move... the 0.4x plugin needs to be moved next,
and then better logging setup in the library if needed, and then
work to add vformat support to btool.