sync: added workarounds for firmware hang issue
This workaround overrides the timeout due to a firmware issue: sometimes
the firmware will hang during a plugin Reconnect(), during the shutdown
stage when the plugin attempts to update the device's dirty flags.
When the device hangs like this, it fails to respond to a Desktop::Open().
To work around this, we set the default timeout to 15 seconds so that
we find this failure early enough to fix it within opensync's 30 second
timeout.  Then if we get such a timeout, we do the Reconnect again and
hope for the best... this often fixes it.

Applied to both plugins.