debian: preliminary effort to build barry+plugins without installing anything
Normal method of building binary packages that depend on others is to
build the library, install the -dev, and build the dependent package.
But it would be nice to be able to build Barry and plugins without
having to install libbarry-dev first.

The old system seemed to do it ok... need to duplicate it in new
system with separate debian directories.