lib: fixed missing GCC visibility bug for libbarrysync library
When a library has its own flags variables in the form of
libbarrysync_la_CXXFLAGS, then these variables must include
AM_CXXFLAGS and AM_CFLAGS as well, otherwise they are not used.

This does not seem to be the case for the linker flags, as far as I
can tell.