lib: added DBName support to parsers and builders
For the upcoming libbarrybackup, it makes sense that parsers and
builders know what database they are working on, since it is
possible they can contain more than one.

This commit adds the DBName argument to SetIds(), which means that
during a Desktop::LoadDatabase(), the database name will automatically
come along with it.

It also adds Builder::GetDBName(), which supports it in the other
direction.