lib: removed the dbId from Builder::Retrieve()
There was never any true connection between the dbId passed into
Retrieve() and the data that it actually did retrieve.  Passing
it in implies that Retrieve() will always have a random-access
database for storage, which is obviously not true.

In the common case, it is Retrieve() that actually determines
what data is returned.  So this commit removes the useless
argument from Retrieve() and changes the template functor
from Id to a reference to Builder itself.  That argument
was only there to differentiate between Parser storage objects
and Builder storage objects.