Add ALX parser
A new API allows to parse a sigle ALX file with OS informations or not.
It allows to parse several ALX files about BlackBerry OS.

Add a new tool balxparse to test the API and dump the result for a
human.

This API uses the XML++ 2.6 library to read XML file. If you haven't the
XML++ library installed, barry will build without the ALX feature.