lib: added Data version of TarFile::ReadNextFile() (optimization)
This is in an effort to reduce data copies in the code, but it
causes code duplication.  It is probably better to turn TarFile into
a complete Barry namespace class someday, using Data objects instead
of std::string, instead of leaving it as a generic reuse library class.

See FIXME notes in the code.