lib: allow Packet class to override receive buffer on the fly
This is for copy-reduction optimization, so that it is possible to
create API's which take Data objects from the application, and
pass them all the way down to the socket level, to avoid copies
wherever possible.

Unfortunately, this API change introduces a slight risk that
the scope of the new receive buffer may be shorter than the scope
of the Packet class itself, but since this should normally only be
used within the library, this is a relatively safe risk.