lib: changed Restore's record retrieval into a state machine
At times, it may be useful to "peek" at the next record's meta
data, so as to make decisions based on DBName, etc.

In order to do this safely, Retrieve() needs to be able to be
called multiple times.  The more fine-grained state machine,
instead of just true/false, is more helpful.