lib: added Connector base class and DesktopConnector class
These classes manage connecting and reconnecting to a device.
Too much technical knowledge was required in the application
in order to successfully connect and reconnect on all devices,
such as in the plugins when refreshing the dirty flags.
Pauses and rescans were needed in some cases.  This knowledge
belongs in the library, not the application.