lib: moved sequence packet checking out of ZeroSocket
This commit moves sequence packet checking out of the ZeroSocket
class and into a standalone function, as well as cleaning up some
other code that does this check manually.

Note that with this commit, HideSequencePacket() support is now
broken, but this is only temporary, to be fixed by further refactoring
coming up.