lib: router: Always queue sequence packets, even if there's a callback
In order to fully implement a SyncSend() style Socket class, we need a
reliable way to handle incoming sequence packets.  The easiest way is
to assume that all sequence packets are queued.  That way, non-sequence
packets are routed to available callbacks, while SyncSend() handling
is still possible via the queued sequence packets.