maint: changed build scripts to allow specifying of all debian targets
The barrychroots and barrylocal configuration scripts can now
specify *all* the debian/rules targets, including 'binary'.
This allows special handling for some distros, using a common
rules makefile.

This may be needed for Debian Squeeze which includes a libgcal
package, version 0.9.5, but it includes the needed fix from 0.9.6.
So binary packages for Debian, but not Ubuntu, can be more lenient.