gui: fixed bug in password prompt in BarryBackup
The Thread class's Connect(password) function did not handle the case
where DeviceInterface::Password() would return false, for low level
USB or protocol errors, and so would assume a connection when one
had not been made.  This would lead eventually to a segfault.

This commit adds the check, and a check in the GUI code to properly
display the error during the connection attempt.