lib: fixed hang in IPModem Close
If the data read thread is still going, then there is no need to do
an explicit BulkRead() in the Close() routine.  Only do this if the
thread is not running.

When two libusb 0.1 reads are done at the same time on the same endpoint
from different threads, it seems like it causes a hang in the pthread_join().
Not sure why.  Did not test with libusb 1.0