os22: filter get_changes based on objtype selection
The 0.22 plugin retrieves changes for calendar and contacts in one
function call, and would filter one or the other based on the Barry
configuration.

But there is another way to filter it, from the opensync side, by
enabling and disabling objtypes.

This commit checks both.

This should not be needed for the 0.4x plugin, since get_changes is
handled in multiple function calls for each objtype.