desktop: added simple EvoDefaultDlg
This dialog lets the user know that the auto-detection worked and
there were sufficient settings to pick a sensible default.  It also
provides a way for the user to configure the paths manually if he so
chooses.

Note: if you are a wxWidgets coder and want to hack on Barry, see this
commit, and send a patch for the missing ICON_INFORMATION icon.
Thanks!