desktop: remove the "file://" futzing code
Upcoming changes will now store the full opensync path config in
memory, regardless if it is a path or a specific evolution item.

By default, the full "file://...." strings will be stored in the config
file and memory now, instead of "file://" in the config file and just
the path in memory.