lib: made Probe configurable to suppress some common USB errors
If Probe cannot communicate with a device initially, this can throw
a Usb::Error exception.  This is not optimal if the device is busy
with another app, or if the device pemissions are incorrect.

This commit makes it configurable for the application to determine
which errors to log instead of throw on.