lib + gui: added Auto Select All checkbox to the backup GUI
If the checkbox is checked, in the backup database list config,
then on every backup, the backup list is taken from the device's
DBDB, not the database list selection.

If you regularly backup everything, this setting prevents you from
ignoring new databases that appear when installing new applications
on your device, etc.