lib: fixed DueDateFlag bug and added AlarmFlag
In modern devices, in the Tasks database, when a DueTime or AlarmTime
are set, respectively, an extra field shows up for each with a flag.

On older devices, such as the 7750, these flags are not available.

This commit adds support for the AlarmFlag.

It also fixes the case where DueDateFlag would be false on old devices,
even though a due time was set.  And similar special casing is done for
the new AlarmFlag.