desktop: Modem: cleaned up pppd permissions checks, and GUI su wrapper
First, the code checking pppd permissions and running tests to make
sure we can execute pppd with the permissions we have were broken.
This commit makes better use of ProgramDetect's values, since we
can't rely on ExecHelper's Run() return for permission status.

Also, not all systems have gksu, so add the ability to set this
at compile time with the --with-guisu compile option.  By default,
it searches for gksu (for Debian), beesu (for Fedora), and kdesu
(for openSUSE) in that order.