Use version strings from configure instead of hardcoded
This commit updates gui, desktop, and both plugins, so each of
their configure.ac scripts contain their own version numbers, which
can be used in code and scripts beneath them.

For gui, and the plugins, it is possible to use autoconf's PACKAGE_VERSION
instead of our own.  For the library, and desktop, the individual
version numbers are needed for library and directory versioning.