desktop: added concept of configured PST type mask
For example, in the Evolution plugin config, it may be difficult for the
end user to find the correct config for Memos or Calendar, if the
autodetect code doesn't work for them on their system.  If the user
just doesn't care about syncing that particular object, it should be
possible to use an empty string, and then disable the corresponding
sync type.

By adding the concept of configured sync types to the plugin classes,
and ANDing the supported sync types with the configured ones, it is
possible to make these configuration possibilities more intuitive
and automatic.