udev: removed udev/69-blackberry.rules which used ACL_MANAGE
According to:
  http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=593881#20

applications should not use ACL_MANAGE since it is not a public
interface.

According to my recent testing on Fedora 16, it seems to work
without this special 69-blackberry.rules too, but could use another
test on Fedora 17.

This commit removes 69-blackberry.rules and all mention of ACL_MANAGE,
and removes the udev file from the rpm spec file.  The Debian packaging
did not use this rule.

Thanks to Bojan Vondra for prompting this research by his openSUSE
spec file patch.