rpm: removed BuildRequires lines from barry.spec
Since we are building with binary-meta, which contains depscripts
for its own dependencies, the BuildRequires lines are just points of
non-portability between Fedora and openSUSE, when built with merely
a call to rpmbuild, like we use.

If we used an official build service, we'd need these, but we don't
yet need them.  Keeping it simple....