desktop: split wxi18n.h and i18n.h more strictly
The GUI code should completely depend on wxWidgets for translation i18n
support, while the library should depend on gettext, or at least a
veneer of gettext, so it is theoretically independent from wxWidgets
for non-gui code.

The _C() macro in wxi18n.h uses a temporary to achieve its results,
and I went through the code where this is affected and double checked
to make sure temporaries work ok.  So far so good.

This commit still assumes that both GUI code and library code use the
same text domain... this may not be proper, since the library should
have its own text domain, so it returns translated strings regardless
of GUI.