desktop: changed PNGButton to use final versions of buttons
Instead of drawing the background, then the button over top of it each time,
during Init(), create the final image, and then during draw, just
draw that image exactly, without transparency.

This also fixes the transparency issues of DrawButtonLabel(), which is
now renamed to DrawButtonLabelDC().

A PNGButton::Erase() member was added, in case button movement is required
in the future.