lib: trim.h: make wrapper function inline, put in namespace, and keep std::
I've run into compile trouble if some functions are not inline, in
headers.

Should keep this wrapper function inside the same namespace as the
trim functions.

The previous code used std::isspace(), so attempt to continue using it.
I don't know if this compiles on WinCE.

Since this changes inline functions definitions, I'm assuming it could
break the ABI, so postpone for 0.19.0.