lib: removing casts where possible
I prefer to use exact variables if possible, instead of casts, in order
to clear compiler warnings.  I hope this still works on Win32.

Reserve casting for actual conversions, such as the cast from float
to time_t in Message2Time(), instead of only hiding compiler warnings.

A compiler warning fixed only with a cast is a compiler warning that
may be incorrectly hidden later.