Fixed incorrect handling of end of QUOTED-PRINTABLE lines
Firstly, changed the name from _fold_lines  to _unfold_lines.

The problem was that =\r\n was being passed through unchanged when the
line is marked QUOTED-PRINTABLE, when in fact it should be a 'soft' line
break. The \r\n should only be left in when not preceeded by the =. It
worked with unix new lines because it's only one character.