lib: make RegisterInterestAndType distinctly different from RegisterInterest
I didn't like exposing RegisterInterest() with a type argument as
the default option, so I modified things a bit and added comments.

Socket::SyncSend() really needs both data and sequence packets in order
to function, so this entire API is really only for the socket opening
race condition.

See the mailing list thread from 2012, with Toby Gray adding WinCE support,
for how this fix was done.