debian: added single-debian-patch in source/options
Barry debian changes are handled in the upstream git repository.
Changes motivated by Debian policy are not always the cause for a new
upstream point release, so are included in patch form here.

This documents that users should look at git for better viewing
of history, instead of looking at the debian patches.